+++ internal/resolvers/betterttv/emote_resolver_test.go
@@ -11,9 +11,9 @@ import (
 	"github.com/Chatterino/api/pkg/config"
 	"github.com/Chatterino/api/pkg/utils"
 	qt "github.com/frankban/quicktest"
-	"github.com/golang/mock/gomock"
 	"github.com/jackc/pgx/v4"
 	"github.com/pashagolub/pgxmock"
+	"go.uber.org/mock/gomock"
 )
 
 func TestEmoteResolver(t *testing.T) {