+++ internal/resolvers/twitch/clip_resolver_test.go
@@ -13,10 +13,10 @@ import (
 	"github.com/Chatterino/api/pkg/config"
 	"github.com/Chatterino/api/pkg/utils"
 	qt "github.com/frankban/quicktest"
-	"github.com/golang/mock/gomock"
 	"github.com/jackc/pgx/v4"
 	"github.com/nicklaw5/helix"
 	"github.com/pashagolub/pgxmock"
+	"go.uber.org/mock/gomock"
 )
 
 func TestClipResolver(t *testing.T) {