+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Minor: Server perks are now sorted alphabetically. (#711)
 - Dev: Added unit tests for the Discord invite resolver. (#711)
+- Dev: Replace github.com/golang/mock with go.uber.org/mock. (#712)
 
 ## 2.0.5